repo: phase 2 - move facedancer future code into package root
