+++ facedancer/magic.py
